refactor(mllp,mllp-transport,mllp-client): extract wire protocol into its own package
Add @glion/mllp-transport, a new package that owns the MLLP wire
protocol primitives — framing constants, encoder, decoder, decoder
stream, dynamic buffer, base MllpError, and TransportError. These
are concerns shared by senders and receivers and have no server
semantics, so they should not live in the server package.

@glion/mllp now depends on @glion/mllp-transport and re-exports the
same symbols from its top-level entry, so existing consumers see no
API change. Internal imports in the server and Node adapter switch to
the new package.

@glion/mllp-client now depends on @glion/mllp-transport instead of
@glion/mllp. Before this change, the client had to pull in the entire
server framework (router, middleware, Mllp class, Node adapter) just
to reach `encode`, `createDecoderStream`, and `MllpError`. The two
packages are now siblings sitting on top of a shared protocol package,
which is what the names imply.

Files moved with `git mv` to preserve history.

Tests: 58 (mllp-transport, moved) + 132 (mllp) + 19 (mllp-ack) +
9 (mllp-client) all green.
