+++ packages/mllp-transport/package.json
@@ -1,15 +1,18 @@
 {
-  "name": "@glion/mllp-client",
+  "name": "@glion/mllp-transport",
   "version": "0.15.3",
-  "description": "MLLP client for sending HL7v2 messages and receiving acknowledgments",
+  "description": "MLLP wire protocol primitives — framing, encoding, decoding, and base error types",
   "keywords": [
-    "client",
+    "decoder",
+    "encoder",
+    "framing",
     "health",
     "healthcare",
     "hl7",
     "hl7v2",
     "mllp",
-    "nodejs",
+    "protocol",
+    "transport",
     "typescript"
   ],
   "homepage": "https://glion.dev",
@@ -21,7 +24,7 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/rethinkhealth/glion.git",
-    "directory": "packages/mllp-client"
+    "directory": "packages/mllp-transport"
   },
   "files": [
     "dist"
@@ -46,18 +49,8 @@
     "test:coverage": "vitest run --coverage",
     "test:watch": "vitest"
   },
-  "dependencies": {
-    "@glion/ack": "workspace:*",
-    "@glion/ast": "workspace:*",
-    "@glion/mllp-transport": "workspace:*",
-    "@glion/parser": "workspace:*",
-    "@glion/util-query": "workspace:*"
-  },
   "devDependencies": {
     "@glion/check-readme": "workspace:*",
-    "@glion/hl7v2": "workspace:*",
-    "@glion/mllp": "workspace:*",
-    "@glion/mllp-ack": "workspace:*",
     "@glion/testing": "workspace:*",
     "@glion/tsconfig": "workspace:*",
     "@types/node": "^25.5.0",
