+++ packages/mllp-transport/src/constants.ts
