+++ packages/mllp-transport/src/decoder-stream.ts
