+++ packages/mllp-transport/src/decoder.ts
