+++ packages/mllp-transport/src/dynamic-buffer.ts
