+++ packages/mllp-transport/src/encoder.ts
