+++ packages/mllp-transport/src/errors.ts
@@ -1,3 +1,5 @@
+import type { TransportErrorCode } from "./types";
+
 /**
  * Options for constructing an MllpError.
  */
@@ -9,9 +11,9 @@ export interface MllpErrorOptions extends ErrorOptions {
 /**
  * Base error class for all MLLP errors.
  *
- * All error types in the MLLP package extend this class.
  * Provides a common `code` field so that any MLLP error can be
  * programmatically identified and translated (e.g., into ACK/NAK responses).
+ * Extended by transport-, server-, and client-side packages.
  */
 export class MllpError extends Error {
   /** Machine-readable error code identifying the error type */
@@ -24,5 +26,21 @@ export class MllpError extends Error {
     this.name = "MllpError";
     this.code = code;
     this.errorCondition = options?.errorCondition;
+  }
+}
+
+/**
+ * Transport-level error for MLLP framing issues (invalid framing, incomplete
+ * messages, etc.)
+ */
+export class TransportError extends MllpError {
+  declare readonly code: TransportErrorCode;
+  /** Position in the data where the error occurred (if applicable) */
+  readonly position?: number;
+
+  constructor(code: TransportErrorCode, message: string, position?: number) {
+    super(code, message);
+    this.name = "TransportError";
+    this.position = position;
   }
 }
