+++ packages/mllp-transport/test/encoder.test.ts
@@ -2,8 +2,8 @@ import {
   MLLP_END_BYTE_1,
   MLLP_END_BYTE_2,
   MLLP_START_BYTE,
-} from "../../src/transport/constants.js";
-import { encode, encodeMultiple } from "../../src/transport/encoder.js";
+} from "../src/constants.js";
+import { encode, encodeMultiple } from "../src/encoder.js";
 
 describe("encode", () => {
   it("should encode a simple ASCII message", () => {
