+++ packages/mllp-transport/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "@glion/tsconfig/library.json",
+  "compilerOptions": { "outDir": "dist", "strictNullChecks": true }
+}
