+++ packages/mllp/package.json
@@ -53,6 +53,7 @@
   },
   "dependencies": {
     "@glion/ast": "workspace:*",
+    "@glion/mllp-transport": "workspace:*",
     "@glion/parser": "workspace:*",
     "@glion/util-query": "workspace:*",
     "unified": "11.0.5",
